gooru/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
autobahn==22.5.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==4.0.0
charset-normalizer==2.0.8
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
Django==3.2.9
django-cors-headers==3.10.0
django-crontab==0.7.1
django-excel==0.0.10
django-filter==21.1
django-multiselectfield==0.1.12
django-oauth-toolkit==1.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.21.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
jwcrypto==1.0
lml==0.1.0
MarkupSafe==2.0.1
msgpack==1.0.4
netaddr==0.8.0
oauthlib==3.1.1
openpyxl==3.0.10
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
pysendpulse==0.1.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-memcached==1.51
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
texttable==1.6.4
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
yookassa==2.3.1
zope.interface==5.4.0
